Read AST from JSON and show code type in TTY
OMG!  The TTY will now show the AST node type in the Solidity pane
header, so for example if the EVM is at the expression `balances[x]'
then the header shows "(IndexAccess)".

That's mildly useful in itself but having access to the AST nodes during
evaluation is really powerful!  I can hardly wait to use it for
cool stuff!