funny/kludgy way of getting the new ethabi
So, there's a new Rust Nix build system in town, but right now I can't
be bothered to make it work with GitHub sources, so I just copied ethabi
into this repository, which is ridiculous, but it's small, so whatever.

There's a problem with installing ethabi on the top level of your user
environment using "nix-env" but we mostly care about it as a dependency
of seth, and for that, it works, so.

I'm in correspondence with the Mr Manager of the new Rust thing, so
hopefully all this will be resolved nicely soon.