`RETURN` opcode handling fixed (in case of variable length return data)
While capturing the returndata, it was the length from the `CALL` opcode,
but not the length from the `RETURN` that was being used as the returndata length.
It means that if a `CALL` was made with `outsize=0`, no returndata got saved
so even subsequent `RETURNDATASIZE` and `RETURNDATACOPY` opcodes could not retrieve it.