Move ethjet into hevm
It was nice to have ethjet as a totally separate library,
but it made it harder to package hevm (e.g. for Hackage),
so I just moved it into hevm, both the C and the binding.

This makes it a little harder to use "cabal build" through Nix because
of linking issues.  Hopefully that can be fixed.