hevm: handle rpc fetch in tty - fixes #257
A minimal fix for #257 where remote contracts are now correctly
fetched when running with the --rpc flag. Attempts to fetch
non-existant contracts will result in a VMFailure.

The fix is somewhat unsatisfactory because the tty interpreter
is not suited to gracefully handling operational failures without a
more extensive refactor of the stepper and other interpreters.

In the meantime rpc fetch is functional, with more robust and
informative error handling to follow.