dapp: use solc binaries directly in dapp --use
This is significantly quicker than using `nix run`. We could have
continued using `nix run`, overriding solc with a static binary: this
would have saved on the download size of the non-static solc being used
previously, but still makes a query to the binary caches and fails
on an offline machine.

Important note: the linux static builds do not have z3 enabled, so SMT
checking will no longer work on linux when using `dapp --use solc:x.y.z`.
To use a z3-enabled solc when using `dapp --use` a user will have to install
it and then use it with either

    dapp --use solc ...

where "solc" is the name of the installed binary, or

    dapp --use /path/to/solc ...

or

    DAPP_SOLC=`which solc` dapp ...

The default solc installed by dapptools (and linked to dapp when invoked
without --use) remains as the nixpkgs z3-enabled version, i.e. we will
still do smt builds with

    dapp build